IE9 file uploads changed to simple file submission, max file size added.

+++ website/templates/project/files.mako
@@ -6,7 +6,8 @@
 %if user_can_edit:
 <div class="container" style="position:relative;">
 ##    <h3 style="max-width: 65%;"><span class="btn btn-success fileinput-button" id="clickable"><i class="icon-plus icon-white"></i><span>Add files...</span></span></h3>
-    <h3 >Drag and drop (or <a href="#" id="clickable">click here</a>) to upload files into <element id="componentName"></element></h3>
+    <h3 id="dropZoneHeader">Drag and drop (or <a href="#" id="clickable">click here</a>) to upload files into <element id="componentName"></element></h3>
+    <div id="fallback"></div>
     <div id="totalProgressActive" style="width: 35%; height: 20px; position: absolute; top: 73px; right: 0;" class>
         <div id="totalProgress" class="progress-bar progress-bar-success" style="width: 0%;"></div>
     </div>
@@ -17,6 +18,18 @@
 <div id="myGrid" class="dropzone files-page"></div>
 </div>
 
+<!--[if lte IE 9]>
+<script>
+    browserComp = false;
+    var htmlString = "    <form action='" + ${info}[0]['uploadUrl'] + "' method='POST' enctype=multipart/form-data>" +
+            "<p><input type=file name=file>" +
+            "<input type=submit value=Upload>" +
+            "<input type='hidden' name='redirect' value='true' />" +
+            "</form>"
+    $('#dropZoneHeader').css('display', 'none');
+    $('#fallback').html(htmlString);
+</script>
+<![endif]-->
 <script>
 
 var extensions = ["3gp", "7z", "ace", "ai", "aif", "aiff", "amr", "asf", "asx", "bat", "bin", "bmp", "bup",
@@ -97,6 +110,11 @@ var Buttons = function(row, cell, value, columnDef, dataContext) {
 
 $('#componentName').text(${info}[0]['name']);
 
+if(typeof(browserComp) === 'undefined'){
+    browserComp = true;
+}
+
+var useDropZone = ${int(user_can_edit)} && browserComp;
 
 var myGrid = HGrid.create({
     container: "#myGrid",
@@ -122,7 +140,7 @@ var myGrid = HGrid.create({
     autoHeight: true,
     forceFitColumns: true,
     largeGuide: false,
-    dropZone: ${int(user_can_edit)},
+    dropZone: useDropZone,
     dropZonePreviewsContainer: false,
     rowHeight: 30,
     navLevel: ${info}[0]['uid'],
@@ -131,7 +149,6 @@ var myGrid = HGrid.create({
     dragToRoot: false,
     dragDrop: false
 });
-
 
 
 myGrid.updateBreadcrumbsBox(myGrid.data[0]['uid']);
@@ -221,17 +238,17 @@ myGrid.hGridOnUpload.subscribe(function(e, args){
     return false;
 });
 
-myGrid.dropZoneObj.options.maxFilesize = 250;
-
-myGrid.dropZoneObj.on("error", function(file, errorMessage, xhr){
-    if(errorMessage.indexOf("Max filesize")!=-1){
-        alert(errorMessage);
-        var item = myGrid.getItemByValue(myGrid.data, file.name, "uid")
-        if(item){
-            myGrid.deleteItems([item['uid']]);
+if(myGrid.dropZoneObj){
+    myGrid.dropZoneObj.on("error", function(file, errorMessage, xhr){
+        if(errorMessage.indexOf("Max filesize")!=-1){
+            alert(errorMessage);
+            var item = myGrid.getItemByValue(myGrid.data, file.name, "uid")
+            if(item){
+                myGrid.deleteItems([item['uid']]);
+            }
         }
-    }
-});
+    });
+}
 
 ##var date1=0;
 ##
